Functionality $5
Gui overhaul for papers: Papers now can be viewed as a category, the
title is clickable and allows to view all info for the single paper.
This can be found in "Example Category". It is scalable. ;View paper
list has been changed to "Administrate Papers" as it is only for admin;
Reject button added for the moderator, however it is not working yet;
Current bugs: Reject button does not pass the correct PaperId,
Validation issues for paper analysis(numerical).